# docs/swagger.yaml
basePath: /api/
definitions:
...
  models.User:
    properties:
      _id:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      password_changed_at:
        type: string
      profile_image:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  services.ChangePasswordRequest:
    properties:
      password:
        type: string
      password_confirmation:
        type: string
    type: object
...
  services.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      profile_image:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  services.CreateUserResponse:
    properties:
      user_id:
        type: string
    type: object
...
  services.GetUserResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
  services.GetUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  services.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      profile_image:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  services.UpdateUserResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
host: localhost:8080
info:
  contact:
    email: maramal@outlook.com
    name: Martin Fernandez
    url: https://mafer.dev
  description: API para administradores de Ayud App
  title: API de Administración de Ayud App
  version: "1.0"
paths:
...
  /admin/users:
    get:
      operationId: get-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.GetUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.GetUsersResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtiene todos los usuarios
    post:
      consumes:
      - application/json
      operationId: create-user
      parameters:
      - description: Datos del usuario
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/services.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.CreateUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Crea un usuario
  /admin/users/{id}:
    delete:
      operationId: delete-user
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Elimina un usuario
    get:
      operationId: get-user
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.GetUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtiene un usuario
    put:
      consumes:
      - application/json
      operationId: update-user
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del usuario
        in: body
        name: UpdateUserRequest
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.UpdateUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Actualiza un usuario
  /admin/users/{id}/password:
    put:
      consumes:
      - application/json
      operationId: change-password
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del usuario
        in: body
        name: ChangePasswordRequest
        required: true
        schema:
          $ref: '#/definitions/services.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Cambia la contraseña de un usuario
  /admin/users/{id}/set-super-admin:
    post:
      operationId: set-super-admin
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Configura un usuario como super administrador
  /admin/users/{id}/unset-super-admin:
    post:
      operationId: unset-super-admin
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Configura un super administrador como usuario
  /admin/users/email/{email}:
    get:
      operationId: get-user-by-email
      parameters:
      - description: Correo electrónico del usuario
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.GetUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtiene un usuario por su correo electrónico
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"